Add RelationField to FieldDefinition, to allow basic relational mapping between objects.
RelationField requires the Repository needed to fetch the Entity via its entityId.

When saving an Entity, both the JSON Converter and HASH Converter will check whether the RelationField has been populated (in other words, whether the Entity behind the entityId has been fetched).
If it has been populated, the entityId will be stored in the property before saving, otherwise it is assumed that the RelationField has not been populated and it is left untouched.